refactor(tui): app.py から描画とレイアウトを切り出す
app.py に画面パーツの描画と Window / Float の組み立てが残っていたため、
app_render.py と app_layout.py に分ける。app.py は state の初期化と
Application の組み立てだけになる。

描画関数はモジュール外から使うので _render_tabs などの _ を外す。
既存の render.py はタブ側が使うテキスト整形ユーティリティなので手を付けず、
app_ プレフィックスで区別する。
