+++ src/redi/tui/app.py
@@ -1,46 +1,21 @@
 import shutil
 from datetime import datetime
-from importlib.metadata import version
 from pathlib import Path
 
 from prompt_toolkit import Application
-from prompt_toolkit.data_structures import Point
 from prompt_toolkit.key_binding import KeyBindings
-from prompt_toolkit.layout import Layout
-from prompt_toolkit.layout.containers import (
-    ConditionalContainer,
-    Float,
-    FloatContainer,
-    HSplit,
-    VSplit,
-    Window,
-)
-from prompt_toolkit.layout.controls import FormattedTextControl
-from prompt_toolkit.utils import get_cwidth
-from prompt_toolkit.widgets import Frame
 
 from redi.api.me import fetch_my_user_id
-from redi.i18n import messages
+from redi.tui.app_layout import build_layout
 from redi.tui.conditions import build_conditions
-from redi.tui.issue.filter_modal import build_filter_float
 from redi.tui.issue.issue_tab import fetch_issues_with_filter, load_journals
 from redi.tui.keybindings import (
     modal_keybindings,
     normal_keybindings,
     submode_keybindings,
 )
-from redi.tui.project_modal import build_project_float
-from redi.tui.state import (
-    FIXED_ROWS,
-    Renderable,
-    TuiPosition,
-    TuiResult,
-    TuiState,
-)
+from redi.tui.state import FIXED_ROWS, TuiPosition, TuiResult, TuiState
 from redi.tui.tabs import TABS
-from redi.tui.time_entry.filter_modal import (
-    build_filter_float as build_time_entry_filter_float,
-)
 
 
 def dump_rendered_screen(app: Application) -> dict:
@@ -74,115 +49,6 @@ def _append_screen_yaml(path: Path, dumped: dict, key: str) -> None:
     )
     with open(path, "a", encoding="utf-8") as f:
         f.write(entry)
-
-
-def _render_tabs(state: TuiState) -> Renderable:
-    parts: Renderable = []
-    for i, (key, tab) in enumerate(TABS.items()):
-        if i > 0:
-            parts.append(("", "  "))
-        style = "reverse" if state.tab == key else ""
-        parts.append((style, f" {tab.label} "))
-    parts.append(("", messages.tui_tab_switch_hint))
-    # 未切替時は名前解決の API を呼ばず config の設定値 (id/identifier) を出す。
-    project = state.project_label or state.effective_project_id()
-    if project:
-        parts.append(
-            (
-                "bold fg:ansicyan",
-                messages.tui_current_project.format(name=project),
-            )
-        )
-    return parts
-
-
-def _render_list_current(state: TuiState) -> Renderable:
-    return TABS[state.tab].render_list(state)
-
-
-def _skip_lines(parts: Renderable, n: int) -> Renderable:
-    """`parts` の先頭から論理行 (newline 区切り) を `n` 個分捨てて残りを返す。
-
-    prompt_toolkit の `Window` は `wrap_lines=True` 時に `get_vertical_scroll`
-    を参照しないため、レンダー結果側で先頭をスライスしてプレビューのスクロール
-    を実現する。
-    """
-    if n <= 0:
-        return list(parts)
-    result: Renderable = []
-    seen = 0
-    started = False
-    for style, text in parts:
-        if started:
-            result.append((style, text))
-            continue
-        nl_in_text = text.count("\n")
-        if seen + nl_in_text < n:
-            seen += nl_in_text
-            continue
-        need = n - seen
-        idx = -1
-        for _ in range(need):
-            idx = text.find("\n", idx + 1)
-        rest = text[idx + 1 :]
-        if rest:
-            result.append((style, rest))
-        started = True
-    return result
-
-
-def _render_preview_current(state: TuiState) -> Renderable:
-    parts = TABS[state.tab].render_preview(state)
-    if state.preview_scroll <= 0:
-        return parts
-    return _skip_lines(parts, state.preview_scroll)
-
-
-def _help_version_label() -> str:
-    return f"redi v{version('redtile')}"
-
-
-def _render_help(state: TuiState) -> Renderable:
-    lines = TABS[state.tab].help_lines
-    width = max(len(key) for key, _ in lines) + 2
-    parts: Renderable = []
-    seen_section = False
-    # バージョン行を右下に寄せるため、本文の最大表示幅 (CJK は2セル) を測る
-    body_width = 0
-    for key, desc in lines:
-        if not desc:
-            if seen_section:
-                parts.append(("", "\n"))
-            parts.append(("bold", f"{key}\n"))
-            seen_section = True
-            body_width = max(body_width, get_cwidth(key))
-        else:
-            parts.append(("bold fg:ansicyan", key.ljust(width)))
-            parts.append(("", f"  {desc}\n"))
-            body_width = max(body_width, width + 2 + get_cwidth(desc))
-    label = _help_version_label()
-    padding = max(0, body_width - get_cwidth(label))
-    parts.append(("", "\n"))
-    parts.append(("fg:ansiwhite", f"{' ' * padding}{label}"))
-    return parts
-
-
-def _render_error_modal(state: TuiState) -> Renderable:
-    body = state.error_modal or ""
-    return [("fg:ansired", body)]
-
-
-def _render_status(state: TuiState) -> Renderable:
-    if state.confirm_delete_prompt is not None:
-        return [("reverse", f" {state.confirm_delete_prompt} ")]
-    if state.flash_message is not None:
-        return [("bold fg:ansiyellow", f" {state.flash_message} ")]
-    if state.search_mode:
-        return [("reverse", f" /{state.search_query}")]
-    hint = TABS[state.tab].status_hint(state)
-    if state.number_buffer:
-        hint = f" [{state.number_buffer}]" + hint
-    return [("reverse", hint)]
 
 
 def run_issue_tui(
@@ -241,108 +107,8 @@ def run_issue_tui(
     modal_keybindings.register(kb, state, conditions)
     submode_keybindings.register(kb, state, conditions)
 
-    preview_window = Window(
-        FormattedTextControl(lambda: _render_preview_current(state)),
-        wrap_lines=True,
-    )
-
-    main_layout = HSplit(
-        [
-            Window(
-                FormattedTextControl(lambda: _render_tabs(state), show_cursor=False),
-                height=1,
-            ),
-            Window(height=1, char="─"),
-            VSplit(
-                [
-                    Window(
-                        FormattedTextControl(
-                            lambda: _render_list_current(state),
-                            show_cursor=False,
-                            get_cursor_position=lambda: Point(
-                                0, TABS[state.tab].get_cursor_y(state)
-                            ),
-                        )
-                    ),
-                    Window(width=1, char="│"),
-                    preview_window,
-                ]
-            ),
-            Window(FormattedTextControl(lambda: _render_status(state)), height=1),
-        ]
-    )
-
-    # Frame を VSplit で挟んで左右に幅1の空白パディングを置く。
-    # Float の真下の行が CJK 文字 (display width=2) で終わると、その2セル目と
-    # Frame の左ボーダーが同じ列に重なり、prompt_toolkit のレンダラが wide
-    # char の幅ぶんカーソルを進めて Frame ボーダーのセルをスキップしてしまう
-    # (= 縁が表示されない)。1セルの空白を挟むとスキップ先がボーダーではなく
-    # 空白セルに変わるので、ボーダーは常に描画される。
-    help_float = Float(
-        content=ConditionalContainer(
-            content=VSplit(
-                [
-                    Window(width=1, char=" "),
-                    Frame(
-                        Window(
-                            FormattedTextControl(
-                                lambda: _render_help(state), show_cursor=False
-                            ),
-                            wrap_lines=False,
-                        ),
-                        title=lambda: messages.tui_help_title.format(
-                            label=TABS[state.tab].label
-                        ),
-                    ),
-                    Window(width=1, char=" "),
-                ]
-            ),
-            filter=conditions.help_modal,
-        ),
-    )
-
-    filter_float = build_filter_float(state, conditions.issue_filter_modal)
-
-    time_entry_filter_float = build_time_entry_filter_float(
-        state, conditions.time_entry_filter_modal
-    )
-
-    project_float = build_project_float(state, conditions.project_modal)
-
-    error_float = Float(
-        content=ConditionalContainer(
-            content=VSplit(
-                [
-                    Window(width=1, char=" "),
-                    Frame(
-                        Window(
-                            FormattedTextControl(
-                                lambda: _render_error_modal(state), show_cursor=False
-                            ),
-                            wrap_lines=True,
-                        ),
-                        title=lambda: messages.tui_error_modal_title,
-                    ),
-                    Window(width=1, char=" "),
-                ]
-            ),
-            filter=conditions.error_modal,
-        ),
-    )
-
     app = Application(
-        layout=Layout(
-            FloatContainer(
-                content=main_layout,
-                floats=[
-                    help_float,
-                    filter_float,
-                    time_entry_filter_float,
-                    project_float,
-                    error_float,
-                ],
-            )
-        ),
+        layout=build_layout(state, conditions),
         key_bindings=kb,
         full_screen=True,
     )
